% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossExpression.R
\name{correlation}
\alias{correlation}
\title{functions
Computes Pearson's correlation between pairs of columns.
If one matrix is provided, the output is the pairwise correlations between its columns.
If two matrices are provided, the output is the pairwise correlations between their columns.}
\usage{
correlation(matrix1, matrix2 = NULL)
}
\arguments{
\item{matrix1}{The first input matrix.}

\item{matrix2}{The second input matrix (optional).}
}
\value{
Returns a correlation matrix.
}
\description{
functions
Computes Pearson's correlation between pairs of columns.
If one matrix is provided, the output is the pairwise correlations between its columns.
If two matrices are provided, the output is the pairwise correlations between their columns.
}
