\name{expression}
\docType{data}
\alias{expression}
\title{Dataset: expression}
\description{
  A cells x genes matrix that represents gene expression data of a mouse brain from a study published in Nature.
}
\format{
  A matrix with X rows (cells) and Y columns (genes):
  \describe{
    \item{column1}{Gene 1 expression values}
    \item{column2}{Gene 2 expression values}
    \item{...}{Further gene expression values}
  }
}
\source{
  The data is derived from the study: \url{https://www.nature.com/articles/s41586-024-07221-6}
}
