% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossExpression.R
\name{get_cooccurrence_stats}
\alias{get_cooccurrence_stats}
\title{Calculates the number of elements common between columns of two matrices.
This function performs a simple dot product when binarize = FALSE.}
\usage{
get_cooccurrence_stats(matrix1, matrix2, sparse = TRUE, binarize = TRUE)
}
\arguments{
\item{matrix1}{The first input matrix.}

\item{matrix2}{The second input matrix.}

\item{sparse}{If TRUE, coerces matrices to sparse format for efficiency.}

\item{binarize}{If TRUE, converts matrices to binary format to count the number of common elements.}
}
\value{
Returns a column by column matrix, where entries represent the number of common elements.
}
\description{
Calculates the number of elements common between columns of two matrices.
This function performs a simple dot product when binarize = FALSE.
}
