\name{locations}
\docType{data}
\alias{locations}
\title{Dataset: locations}
\description{
  A dataset containing the spatial coordinates (x, y) of cells in a mouse brain derived from a study published in Nature.
}
\format{
  A data frame with X rows (cells) and 2 columns (x, y coordinates):
  \describe{
    \item{pos_x}{The x-coordinate representing the horizontal position of the cell}
    \item{pos_y}{The y-coordinate representing the vertical position of the cell}
  }
}
\source{
  The data is derived from the study: \url{https://www.nature.com/articles/s41586-024-07221-6}
}