% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossExpression.R
\name{scale_01}
\alias{scale_01}
\title{Scales a vector between 0 and 1, including for negative values.}
\usage{
scale_01(vector)
}
\arguments{
\item{vector}{The input vector to be scaled.}
}
\value{
Returns a vector scaled between 0 and 1.
}
\description{
Scales a vector between 0 and 1, including for negative values.
}
